requirements.txt
discord.py==2.3.2
gspread==4.0.1
oauth2client==4.1.3
python-dotenv==1.0.1
